Sorted out the mess of validation files for contributions, moved all into contributions helper.
