+++ assets/pages/oneoff-contributions/actions/oneoffContributionsActions.js
@@ -4,7 +4,7 @@
 
 import { setStripeAmount } from 'helpers/stripeCheckout/stripeCheckoutActions';
 import { setPayPalExpressAmount } from 'helpers/payPalExpressCheckout/payPalExpressCheckoutActions';
-import validateContribution from 'helpers/validation';
+import { validate as validateContribution } from 'helpers/contributions';
 
 
 // ----- Types ----- //
